chore: add basic shortcuts infrastructure
This commit adds the basic infrastructure for handling shortcuts in the
extension and the preferences dialog. It includes a new utility singleton
to manage shortcuts, new prefs components to expose shortcuts in the
preferences dialog. This also supports the idea of multi-stage shortcuts
(See [1] for more details).

[1]: https://leleat.github.io/posts/keyboard-shortcut-sequences-in-gnome-extensions/

Part-of: https://github.com/Leleat/floating-scroll/pull/6