ci: initialize npm project before using commitlint
commitlint doesn't seem to play well with deno. I am getting an error
when trying to run the commitlint CLI command because it is looking for
another commitlint package in the project directory. So let's just
initialize an npm project before installing the commitlint packages in
the CI to workaround it easily.

fixes: 38861f60382d6d35486c7836e0eae995198daba8
Part-of: https://github.com/Leleat/floating-scroll/pull/6