chore: switch from deno to node/npm
I tried to go with deno for this project because I thought it was nice
how it comes with a lot of stuff bundled in making external dependencies
like a formatter, linter, and test runner unnecessary. However, now I'm
going to switch back to node/npm for a few reasons:

- When I was writing tests, I found the documentation to be lacking. It
  was easier to use the jest-like syntax and looking up jest's docs. At
  first that was fine when the code was "purely functional" and didn't
  interact with GNOME's APIs. But now the modules are getting bigger and
  need to (indirectly) call at least some GNOME APIs. I couldn't find
  out how to mock it in deno. So let's switch to deno.
- I prefer prettier over deno's builtin formatter. deno's formatter can
  be harder to read. For instance, instead of breaking up an argument
  list for a function call after each argument and putting each argument
  on its own line, it may break at a curly brace in an object, so it's
  hard to determine where an argument begins and ends. Also, deno's
  formatter can be kinda circumvented by manual line breaks meaning that
  it may not be as consistent as prettier.

So in the end I am introducing some external dependencies again, making
the switch to deno kinda pointless. So let's go with node and npm again.

Part-of: https://github.com/Leleat/floating-scroll/pull/7