chore: add initial version of core data model
This commit adds the initial version of the core data model with some
basic settings and shortcuts to test the behavior... and having tested
the behavior a bit in practice, I am not sure if the column based design
works with floating windows. Maybe the windows should all fill the whole
column... Whatever, let's decide that for later.

The core data model is in src/shell/core and consists of the following

- workspaceView: This is the intersection to the user. It will create
  animations of the windows and pass on the user input to our model.
- eventSystem: This will be instanced by the workspaceView to aggregate
  and track the events that are happening. This might be useful in the
  future for debugging, if we want to understand the sequence of actions
  a user took to get to a certain state.
- workspaceModel: This is the core data model that represents the state
  of a workspace and its windows. A workspace model corresponds to a
  single workspace and will be made up of a row with an infinite number
  of columns of floating windows.

For now the model has not been fully unit-tested. I've written some
initial tests when I was still using deno. But now that I switched the
tests are broken and need to be ported to vitest. Do that in a follow-up
pull request.

Part-of: https://github.com/Leleat/floating-scroll/pull/8