fix: workspaceModel shouldn't use the window rect
The previous tests did uncover a bug when inserting windows. The methods
relied on window's frames which shouldn't be the case since the window
frame may not be the same as a columns since floating-scroll doesn't
have that much control over the window's frame. So use the column's rect
i.e. the intended rect for the window.

Additionally, this commits contains some minor changes like "copying" an
array, and being stricter on an assert.

Part-of: https://github.com/Leleat/floating-scroll/pull/9